Don't reuse identity
It is important that the Tor data directory is maintained between
rebuilds of of the image because otherwise a new one will be generated
and you will get a new fingerprint. This is bad because the Tor network
takes stats on all relay. Furthermore, there is a 2 week "low-traffic"
grace period.